+++ joystick/graph.py
@@ -113,15 +113,6 @@ class Graph(Frame):
         self.ax.plot(0, 0, kwargs.pop('fmt'), **core.linekwargs(kwargs))
         self._scale_axes(force=True)
         self._callmthd(after, **kwargs)
-        # @@@ remove that soon
-        # core.INITMETHOD left for backward compatibility
-        if core.INITMETHOD not in after \
-            and core.INITMETHOD not in before \
-            and hasattr(self, core.INITMETHOD):
-            print("DEPRECATION WARNING: You should add the decorator " \
-                  "`@_callit('after', 'init')` on `{}`. Refer to example.py" \
-                  " ".format(core.INITMETHOD))
-            self._callmthd(core.INITMETHOD, **kwargs)
 
     def _init_basic_graph(self, **kwargs):
         self._xylim = tuple(kwargs.pop('xylim')[:4])