+++ joystick/joystick.py
@@ -47,15 +47,6 @@ class Joystick(object):
         self._frames = []
         self._running = False
         self._callmthd(after, **kwargs)
-        # @@@ remove that soon
-        # core.INITMETHOD left for backward compatibility
-        if core.INITMETHOD not in after \
-            and core.INITMETHOD not in before \
-            and hasattr(self, core.INITMETHOD):
-            print("DEPRECATION WARNING: You should add the decorator " \
-                  "`@callit('after', 'init')` on `{}`. Refer to example.py" \
-                  " ".format(core.INITMETHOD))
-            self._callmthd(core.INITMETHOD, **kwargs)
 
     _extract_callit = core.extract_callit
     _get_infinite_loop_fcts = core.get_infinite_loop_fcts